Memperbaiki masalah

+++ src/main/java/org/jarvis/leave/repository/DetailPengajuanCutiRepository.java
@@ -1,16 +1,14 @@
 package org.jarvis.leave.repository;
 
 import org.jarvis.leave.model.DetailPengajuanCuti;
-import org.jarvis.leave.model.Employee;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
 @Repository
-public interface DetailPengajuanCutiRepository extends JpaRepository<DetailPengajuanCuti, Integer> {
+public interface DetailPengajuanCutiRepository extends JpaRepository<DetailPengajuanCuti, Long> {
 
     @Override
     @Query("select u from DetailPengajuanCuti u where u.isDeleted=false")
