+++ src/main/java/org/jarvis/leave/repository/JenisCutiRepository.java
@@ -3,13 +3,12 @@ package org.jarvis.leave.repository;
 import org.jarvis.leave.model.JenisCuti;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
 @Repository
-public interface JenisCutiRepository extends JpaRepository<JenisCuti, Integer> {
+public interface JenisCutiRepository extends JpaRepository<JenisCuti, Long> {
 
     @Override
     @Query("select u from JenisCuti u where u.isDeleted=false")
